[ScriptTracker] Add separate Log substack
Split out the log into a separate stack which allows the main stack to be smaller and the log to be hidden.  The log opens via a button as a palette so that closing it does not prompt for a save.  Buttons on the main stack allow going between toplevel and palette.

Performed some line wrapping to keep lines under 90 characters