[Stack Tracker] Add script counter field
As a stack is evaluated, the current count of script objects is updated.  Objects with no scripts or behaviors are not counted, just objects that would generate a file.  Count is updated whether the file needs to be exported or not.  The log will contain the names of the files that were exported or imported.