[Stack Tracker] Better detection of file changes
Changed the automatic update logic to only import files that actually changed instead of iterating over the entire stack.  An array is stored with file times and when the directory data changes, only files with a newer update time are imported.  This avoids trying to update when the access time was updated as before.

When a manual sync is done, automatic is turned off.  This is due to the need to reset the stored data.

When a new stack is selected, automatic is turned off since the previous directory and information is no longer valid.

Stack list is now filtered for only those user stacks with a mode > 0.