[Script Tracker] Remove extra files after export
After a full sync, any extra script files are moved to the extra folder (`removeExtraFiles`).  They are renamed prepended with the seconds to avoid possible name conflicts.  New constant added for default extra folder location.

Factored out the folder creation logic to a single handler that is used for all 3 folders that are needed (`ensureDirectoryExists`).

Added `exportMainStackScripts` to remove the recursive calls and make removing extra files cleaner.  Split the logic so that what happens for each stack remains in `exportStackScripts`.  Diff and extra file handling is done in the new handler.

Moved modification time data to the xDataA array, times are captured when exporting scripts

When performing a manual sync, auto remains enabled