[Script Tracker] manually create long names and ids
In order to avoid some issues where long ids and long names can vary, manually create the values that are used in the script exports.  If a background is on a card that has not been visited, then it is reported as "bkgnd...".  If a card that contains it has been visited, then it is reported in the form "object... of card..." for the last card visited.  Objects within the shared group are reported similarly.  If an object is on a card that has not been visited, then the same issue is noted:  "object... of stack..." vice "object... of card... of stack...".

A previous update did this for the background/shared groups, but it did not carry through to the objects within the group.

Since the long id/name is now created manually, moved the shortening of the stack name to the creation of the long id/name.

Added a line for the widget kind to the header of exported scripts

Added name to card id 1002 ("Tracker")