[Script Tracker] Add external editor preference
This update enables a new External Editor function.  Using a subscription to the "ideEditScript" message, requests to edit scripts of the tracked stack are now processed.  Any object within the current main stack will trigger the external editor if the Automatic File Import is enabled and the Edit scripts in external editor preference is enabled.

The default setting for the external editor is the editor used for LCB files (set via the Extension Builder).  A button is provided to allow selection of the application easily.

Created a card in the main stack for the "ScriptTrackerBehavior" button.  Also moved the "P" and "T" buttons used in development to quickly Top level or Palette the visible stacks in the project.

Ensured that when the stack name is changed that the pref window also changed to the stack tab.  Also clear the count field.

Removed generation of empty files just to identify objects with behaviors set.