[NestedDGBehavior] Demo of using a behavior with a DG
The current release of the LC DG does not work with a behavior inserted
between the DG group and the behavior.  This stack is a demo of how that
can be done now while waiting for the code to be incorporated into the
release version.  This change is simpler than what has been submitted.

This stack will only work on LC 9 since the DG was scriptified.  The function
would work in LC 8, but the code to insert it would be different since it
is in the actual button of the stack.