[ScriptTracker] Add script index file
When a full sync is performed, a script index file will be created.
It will reside one level above the script exports - normally in the same
folder as the main stack file.  The name will be the same as the main
stack file with and `.md` extension.  The long name of the object will
be the heading followed by the long ID, behaviors, and a link to the
file.